refactor(bcl): update import paths for services and entities to improve module organization and maintainability
